+++ .idea/misc.xml
@@ -3,17 +3,14 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
-
+        <entry key="../../../../../layout/custom_preview.xml" value="0.1" />
         <entry key="..\:/Users/Daniel/DHBW/5_Semester/Projekt/App/App_zur_Aussagenlogik/app/src/main/res/layout/activity_main.xml" value="0.1921875" />
         <entry key="..\:/Users/Daniel/DHBW/5_Semester/Projekt/App/App_zur_Aussagenlogik/app/src/main/res/layout/fragment_formeln_input.xml" value="0.13020833333333334" />
         <entry key="..\:/Users/Daniel/DHBW/5_Semester/Projekt/App/App_zur_Aussagenlogik/app/src/main/res/layout/fragment_item.xml" value="0.1921875" />
         <entry key="..\:/Users/Daniel/DHBW/5_Semester/Projekt/App/App_zur_Aussagenlogik/app/src/main/res/layout/fragment_resolution.xml" value="0.19067028985507245" />
         <entry key="..\:/Users/Daniel/DHBW/5_Semester/Projekt/App/App_zur_Aussagenlogik/app/src/main/res/layout/fragment_tableaux.xml" value="0.1921875" />
-
-        <entry key="../../../../../layout/custom_preview.xml" value="0.1" />
         <entry key="..\:/Users/demayela/AndroidStudioProjects/App_zur_Aussagenlogik/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.26615384615384613" />
         <entry key="..\:/Users/demayela/AndroidStudioProjects/App_zur_Aussagenlogik/app/src/main/res/drawable/ic_launcher_background.xml" value="0.26615384615384613" />
-
         <entry key="..\:/Users/demayela/AndroidStudioProjects/App_zur_Aussagenlogik/app/src/main/res/layout/activity_main.xml" value="0.264" />
         <entry key="..\:/Users/demayela/AndroidStudioProjects/App_zur_Aussagenlogik/app/src/main/res/layout/fragment_aboutus.xml" value="0.20606884057971014" />
         <entry key="..\:/Users/demayela/AndroidStudioProjects/App_zur_Aussagenlogik/app/src/main/res/layout/fragment_main.xml" value="0.2" />
@@ -35,6 +32,7 @@
         <entry key="app/src/main/res/layout/fragment_normalform.xml" value="0.30027173913043476" />
         <entry key="app/src/main/res/layout/fragment_resolution.xml" value="0.3567708333333333" />
         <entry key="app/src/main/res/layout/fragment_tableaux.xml" value="0.3567708333333333" />
+        <entry key="app/src/main/res/layout/fragment_zwei_formel.xml" value="0.3567708333333333" />
         <entry key="app/src/main/res/layout/history_row_layout.xml" value="0.1" />
         <entry key="app/src/main/res/menu/bottom_nav_menu.xml" value="0.3567708333333333" />
       </map>
