+++ app/src/main/java/com/dhbw/app_zur_aussagenlogik/core/ParserException.java
@@ -3,6 +3,8 @@ package com.dhbw.app_zur_aussagenlogik.core;
 public class ParserException extends Exception{
 
     private int fehlercode;
+    private int[][] truthTable;
+
 
     public ParserException(int fehlercode){
         this.fehlercode = fehlercode;
@@ -14,5 +16,13 @@ public class ParserException extends Exception{
 
     public void setFehlercode(int fehlercode){
         this.fehlercode = fehlercode;
+    }
+
+    public int[][] getTruthTable() {
+        return truthTable;
+    }
+
+    public void setTruthTable(int[][] truthTable) {
+        this.truthTable = truthTable;
     }
 }
