--- app/src/main/java/com/dhbw/app_zur_aussagenlogik/core/Wertetabelle.java
@@ -1,4 +1,0 @@
-package com.dhbw.app_zur_aussagenlogik.core;
-
-public class Wertetabelle {
-}
