+++ app/src/main/res/values/strings.xml
@@ -22,5 +22,7 @@
     <string name="anleitung">Anleitung</string>
     <string name="aboutUs">Über uns</string>
     <string name="historyMenu">Verlauf</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 
 </resources>
